introduce HasUniqueStringIds
