+++ src/Illuminate/Database/Eloquent/Concerns/HasUlids.php
@@ -7,25 +7,7 @@ use Illuminate\Support\Str;
 
 trait HasUlids
 {
-    /**
-     * Initialize the trait.
-     *
-     * @return void
-     */
-    public function initializeHasUlids()
-    {
-        $this->usesUniqueIds = true;
-    }
-
-    /**
-     * Get the columns that should receive a unique identifier.
-     *
-     * @return array
-     */
-    public function uniqueIds()
-    {
-        return [$this->getKeyName()];
-    }
+    use HasUniqueStringIds;
 
     /**
      * Generate a new ULID for the model.
@@ -38,53 +20,13 @@ trait HasUlids
     }
 
     /**
-     * Retrieve the model for a bound value.
+     * Determine if value is a valid ULID.
      *
-     * @param  \Illuminate\Database\Eloquent\Model|\Illuminate\Database\Eloquent\Relations\Relation<*, *, *>  $query
      * @param  mixed  $value
-     * @param  string|null  $field
-     * @return \Illuminate\Contracts\Database\Eloquent\Builder
-     *
-     * @throws \Illuminate\Database\Eloquent\ModelNotFoundException
-     */
-    public function resolveRouteBindingQuery($query, $value, $field = null)
-    {
-        if ($field && in_array($field, $this->uniqueIds()) && ! Str::isUlid($value)) {
-            throw (new ModelNotFoundException)->setModel(get_class($this), $value);
-        }
-
-        if (! $field && in_array($this->getRouteKeyName(), $this->uniqueIds()) && ! Str::isUlid($value)) {
-            throw (new ModelNotFoundException)->setModel(get_class($this), $value);
-        }
-
-        return parent::resolveRouteBindingQuery($query, $value, $field);
-    }
-
-    /**
-     * Get the auto-incrementing key type.
-     *
-     * @return string
-     */
-    public function getKeyType()
-    {
-        if (in_array($this->getKeyName(), $this->uniqueIds())) {
-            return 'string';
-        }
-
-        return $this->keyType;
-    }
-
-    /**
-     * Get the value indicating whether the IDs are incrementing.
-     *
      * @return bool
      */
-    public function getIncrementing()
+    protected function isValidKey($value): bool
     {
-        if (in_array($this->getKeyName(), $this->uniqueIds())) {
-            return false;
-        }
-
-        return $this->incrementing;
+        return Str::isUlid($value);
     }
 }
